refactor(server): extract boot modes and sidecar loaders
Move punctuation / ITN modes and thread budgeting into
boot/modes.rs and punct / VAD / hotword / INT8 loaders into
boot/sidecars.rs. Public gigastt::boot::* paths are unchanged.
